# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vieva Connect API",
    "version": "1.0.0",
    "description": "API REST pour l'application Vieva Connect - Une plateforme de communication familiale et de suivi de proximité avec alertes d'urgence pour personnes âgées.",
    "contact": {
      "name": "Vieva Connect Support",
      "email": "support@vievaconnect.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Serveur de développement (HTTP)"
    },
    {
      "url": "https://vievaconnectbackend.vievaconnect.com/api",
      "description": "Serveur de production (HTTPS)"
    }
  ],
...
  "paths": {
...
    "/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Connexion utilisateur",
        "description": "Authentifie un utilisateur et retourne un token JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "mdp"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Adresse email de l'utilisateur",
                    "example": "jean.dupont@example.com"
                  },
                  "mdp": {
                    "type": "string",
                    "format": "password",
                    "minLength": 1,
                    "description": "Mot de passe de l'utilisateur",
                    "example": "motdepasse123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connexion réussie",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "Token JWT pour les requêtes authentifiées",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    },
                    "expirationTime": {
                      "type": "number",
                      "description": "Timestamp d'expiration du token",
                      "example": 1735689600000
                    },
                    "personne": {
                      "$ref": "#/components/schemas/UserPublic"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Utilisateur non trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Mot de passe invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Erreur de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/create": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Créer un nouveau compte utilisateur",
        "description": "Enregistre un nouvel utilisateur et retourne un token JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["nom", "prenom", "adresse", "telephone", "email", "mdp"],
                "properties": {
                  "nom": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 30,
                    "description": "Nom de famille",
                    "example": "Dupont"
                  },
                  "prenom": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 30,
                    "description": "Prénom",
                    "example": "Jean"
                  },
                  "adresse": {
                    "type": "string",
                    "minLength": 10,
                    "description": "Adresse complète",
                    "example": "123 Rue de la Paix, 75001 Paris"
                  },
                  "telephone": {
                    "type": "string",
                    "minLength": 10,
                    "description": "Numéro de téléphone",
                    "example": "0612345678"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Adresse email unique",
                    "example": "jean.dupont@example.com"
                  },
                  "mdp": {
                    "type": "string",
                    "format": "password",
                    "minLength": 8,
                    "description": "Mot de passe (minimum 8 caractères)",
                    "example": "MotDePasse123!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Utilisateur créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "Token JWT"
                    },
                    "personne": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "L'utilisateur existe déjà",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Erreur de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/profile/{userId}": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Obtenir le profil d'un utilisateur",
        "description": "Récupère les informations détaillées d'un utilisateur",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "objectid"
            },
            "description": "ID de l'utilisateur"
          }
        ],
        "responses": {
          "200": {
            "description": "Profil récupéré avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Utilisateur non trouvé"
          }
        }
      },
      "put": {
        "tags": ["Authentication"],
        "summary": "Modifier le profil utilisateur",
        "description": "Met à jour les informations du profil de l'utilisateur connecté",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de l'utilisateur (doit correspondre à l'utilisateur connecté)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nom": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 30,
                    "description": "Nouveau nom"
                  },
                  "prenom": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 30,
                    "description": "Nouveau prénom"
                  },
                  "adresse": {
                    "type": "string",
                    "minLength": 10,
                    "description": "Nouvelle adresse"
                  },
                  "telephone": {
                    "type": "string",
                    "minLength": 10,
                    "description": "Nouveau numéro de téléphone"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Nouvelle adresse email"
                  },
                  "mdp": {
                    "type": "string",
                    "minLength": 8,
                    "description": "Nouveau mot de passe"
                  },
                  "ancienMdp": {
                    "type": "string",
                    "description": "Ancien mot de passe (requis si mdp est fourni)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profil modifié avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Profil modifié avec succès"
                    },
                    "personne": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erreur de validation ou ancien mot de passe incorrect"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Vous ne pouvez modifier que votre propre profil"
          },
          "409": {
            "description": "Cette adresse email est déjà utilisée"
          }
        }
      }
    },
    "/verifyToken": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Vérifier la validité d'un token JWT",
        "description": "Vérifie si un token JWT est toujours valide",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token"],
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "Token JWT à vérifier"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token valide",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token invalide"
          }
        }
      }
    },
    "/verify_user": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Vérifier l'utilisateur pour ConnectyCube",
        "description": "Vérifie le token et retourne l'ID utilisateur pour ConnectyCube",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token"],
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "Token JWT"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Utilisateur vérifié",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "ID de l'utilisateur"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token invalide"
          }
        }
      }
    },
...
    "/delete-account": {
      "delete": {
        "tags": ["Authentication"],
        "summary": "Supprimer le compte utilisateur",
        "description": "Supprime définitivement le compte de l'utilisateur connecté et anonymise toutes ses données personnelles. Cette action est irréversible et supprime/anonymise : les données personnelles, les contacts, les listes de notifications, les alertes (annulées), les familles créées, et retire l'utilisateur des conversations et familles dont il est membre.",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "Compte supprimé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Votre compte a été supprimé avec succès. Toutes vos données personnelles ont été anonymisées."
                    },
                    "deletionSummary": {
                      "type": "object",
                      "properties": {
                        "contactsDeleted": {
                          "type": "integer",
                          "description": "Nombre de contacts supprimés"
                        },
                        "notificationListsDeleted": {
                          "type": "integer",
                          "description": "Nombre de listes de notifications supprimées"
                        },
                        "alertsCancelled": {
                          "type": "integer",
                          "description": "Nombre d'alertes annulées"
                        },
                        "familiesDeleted": {
                          "type": "integer",
                          "description": "Nombre de familles supprimées (créées par l'utilisateur)"
                        },
                        "familyMembershipsRemoved": {
                          "type": "integer",
                          "description": "Nombre d'appartenances familiales retirées"
                        },
                        "conversationsUpdated": {
                          "type": "integer",
                          "description": "Nombre de conversations mises à jour"
                        },
                        "emptyConversationsDeleted": {
                          "type": "integer",
                          "description": "Nombre de conversations vides supprimées"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Le compte a déjà été supprimé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Utilisateur non trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erreur serveur lors de la suppression",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenu via l'endpoint /login"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "Token JWT stocké dans un cookie"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "objectid",
            "description": "ID unique de l'utilisateur"
          },
          "nom": {
            "type": "string",
            "minLength": 2,
            "maxLength": 30,
            "description": "Nom de famille"
          },
          "prenom": {
            "type": "string",
            "minLength": 2,
            "maxLength": 30,
            "description": "Prénom"
          },
          "adresse": {
            "type": "string",
            "description": "Adresse postale"
          },
          "telephone": {
            "type": "string",
            "description": "Numéro de téléphone"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Adresse email unique"
          },
          "role": {
            "type": "string",
            "description": "Rôle de l'utilisateur",
            "example": "parent"
          },
          "fcmToken": {
            "type": "string",
            "nullable": true,
            "description": "Token Firebase Cloud Messaging"
          },
          "fcmTokenUpdatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Date de dernière mise à jour du token FCM"
          },
          "deviceInfo": {
            "type": "object",
            "properties": {
              "platform": {
                "type": "string",
                "enum": ["ios", "android", "web"]
              },
              "version": {
                "type": "string"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date de création du compte"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date de dernière modification"
          }
        },
        "required": ["nom", "prenom", "email"]
      },
      "UserPublic": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "objectid"
          },
          "nom": {
            "type": "string"
          },
          "prenom": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "telephone": {
            "type": "string"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Message d'erreur"
          }
        },
        "required": ["error"]
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Non autorisé - Token JWT manquant ou invalide",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Unauthorized"
            }
          }
        }
      }
    }
  },
  "security": []
}